Renamed file and cleaned up check
