--- frontend/src/framework/internal/utils/tanstackCacheTime.ts
@@ -1,14 +1,0 @@
-const DEBUG_FLAG_PREFIX = "webviz-debug:";
-
-/**
- * Passes a given time, unless the `disable-tanstack-cache` debug flag is true
- * @param time time (ms) that the cache is valid for
- * @returns
- */
-export function tanstackCacheTime(time: number) {
-    const cacheFlag = localStorage.getItem(DEBUG_FLAG_PREFIX + "disable-tanstack-cache");
-
-    if (["true", true, 1].includes(cacheFlag as any)) return 0;
-
-    return time;
-}
